Improve mock data across all components in preview
