+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ui/chat/components/MessageListView.kt
@@ -18,6 +18,8 @@ import androidx.compose.ui.unit.dp
 import com.azure.android.communication.ui.chat.models.MessageInfoModel
 import com.azure.android.communication.ui.chat.presentation.ui.viewmodel.MessageViewModel
 import com.azure.android.communication.ui.chat.presentation.ui.viewmodel.toViewModelList
+import com.azure.android.communication.ui.chat.preview.MOCK_LOCAL_USER_ID
+import com.azure.android.communication.ui.chat.preview.MOCK_MESSAGES
 import com.azure.android.communication.ui.chat.service.sdk.wrapper.ChatMessageType
 import com.azure.android.communication.ui.chat.utilities.outOfViewItemCount
 import kotlinx.coroutines.launch
@@ -54,29 +56,7 @@ internal fun MessageListView(
 internal fun PreviewMessageListView() {
     MessageListView(
         modifier = Modifier.padding(0.dp),
-        messages = listOf(
-            MessageInfoModel(
-                messageType = ChatMessageType.TEXT,
-                content = "Test Message",
-                internalId = null,
-                id = null
-            ),
-
-            MessageInfoModel(
-                messageType = ChatMessageType.TEXT,
-                content = "Test Message 2 ",
-                internalId = null,
-                id = null
-
-            ),
-            MessageInfoModel(
-                messageType = ChatMessageType.TEXT,
-                content = "Test Message 3",
-                internalId = null,
-                id = null
-
-            ),
-        ).toViewModelList(""),
+        messages = MOCK_MESSAGES.toViewModelList(MOCK_LOCAL_USER_ID),
         scrollState = LazyListState(),
     )
 }
